Day 11 Part 2 solution
Replace an unnecessary enormous loop with using the number directly.
This makes the solution run almost instantly!
Key change is to use a dictionary of number of time a stone exists,
rather than a list that maintains the order with lots of duplicated.
This means we don't run out of memory, and have fewer discrete items to
iterate over.